BuildMaster: Cleanup builders. Untangle spaghetti
